# contract/openapi.yaml
openapi: 3.1.0
info:
  title: Identity Service
  version: 2.0.0
  description: >
    Central identity and access management service.
    Handles member registration, authentication (email/password + social OAuth),
    legacy member migration, session management, token verification,
    organisational hierarchy, group management, and permissions.
servers:
  - url: http://localhost:3000
    description: Local development
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:      { type: string }
            message:   { type: string }
            requestId: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field:   { type: string }
                  message: { type: string }
                required: [message]
          required: [code, message, requestId, details]
      required: [error]
...
    Group:
      type: object
      properties:
        id:           { type: string }
        name:         { type: string }
        description:  { type: string, nullable: true }
        is_active:    { type: boolean }
        member_count: { type: integer }
        created_at:   { type: string, format: date-time }
      required: [id, name, is_active, created_at]
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT issued by the identity service.
        Obtain via /v1/auth/register, /v1/auth/login, or OAuth flows.
paths:
...
  /v1/groups/interest:
    post:
      summary: Create interest group
      operationId: createInterestGroup
      tags: [Groups]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:        { type: string, minLength: 1 }
                description: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Group' }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: Insufficient scope, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: Name exists, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    get:
      summary: List interest groups
      operationId: listInterestGroups
      tags: [Groups]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: includeInactive, in: query, required: false, schema: { type: string, enum: ['true', 'false'] } }]
      responses:
        '200':
          description: List with member counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Group' } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/groups/interest/{id}:
    get:
      summary: Get interest group by ID
      operationId: getInterestGroup
      tags: [Groups]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: Group with member count
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Group' }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    patch:
      summary: Update interest group
      operationId: updateInterestGroup
      tags: [Groups]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:        { type: string, minLength: 1 }
                description: { type: string }
                isActive:    { type: boolean }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Group' }
        '400': { description: No fields, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: Insufficient scope, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    delete:
      summary: Deactivate interest group
      operationId: deleteInterestGroup
      tags: [Groups]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '204': { description: Deactivated }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: Insufficient scope, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/groups/affinity:
    post:
      summary: Create affinity group
      operationId: createAffinityGroup
      tags: [Groups]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:        { type: string, minLength: 1 }
                description: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Group' }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: Insufficient scope, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: Name exists, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    get:
      summary: List affinity groups
      operationId: listAffinityGroups
      tags: [Groups]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: includeInactive, in: query, required: false, schema: { type: string, enum: ['true', 'false'] } }]
      responses:
        '200':
          description: List with member counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Group' } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /v1/groups/affinity/{id}:
    get:
      summary: Get affinity group by ID
      operationId: getAffinityGroup
      tags: [Groups]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: Group with member count
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Group' }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    patch:
      summary: Update affinity group
      operationId: updateAffinityGroup
      tags: [Groups]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:        { type: string, minLength: 1 }
                description: { type: string }
                isActive:    { type: boolean }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Group' }
        '400': { description: No fields, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: Insufficient scope, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    delete:
      summary: Deactivate affinity group
      operationId: deleteAffinityGroup
      tags: [Groups]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '204': { description: Deactivated }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '403': { description: Insufficient scope, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...